+++ client/views/posts/post_page.js
@@ -14,11 +14,7 @@ Template[getTemplate('post_page')].helpers({
 })
 
 Template[getTemplate('post_page')].rendered = function(){
-  // if((scrollToCommentId=Session.get('scrollToCommentId')) && !this.rendered && $('#'+scrollToCommentId).exists()){
-  //   scrollPageTo('#'+scrollToCommentId);
-  //   Session.set('scrollToCommentId', null);
-  //   this.rendered=true;
-  // }
+  $('body').scrollTop(0);
   if(this.data) // XXX
     document.title = $(".post-title").text();
 }