+++ packages/telescope-search/lib/client/routes.js
@@ -3,7 +3,12 @@ adminNav.push({
   label: 'Search Logs'
 });
 
+
 Meteor.startup(function () {
+
+  PostsSearchController = PostsListController.extend({
+    view: 'search'
+  });
 
   Router.onBeforeAction(Router._filters.isAdmin, {only: ['logs']});
 
@@ -13,7 +18,7 @@ Meteor.startup(function () {
 
     this.route('search', {
       path: '/search/:limit?',
-      controller: PostsListController    
+      controller: PostsSearchController    
     });
 
     // Search Logs